+++ packages/stg-evm-v2/devtools/config/mainnet/01/asset.eth.config.ts
@@ -23,7 +23,7 @@ export default async (): Promise<OmniGraphHardhat<AssetNodeConfig, AssetEdgeConf
     const arbPoint = getAssetPoint(EndpointId.ARBITRUM_V2_MAINNET)
     const basePoint = getAssetPoint(EndpointId.BASE_V2_MAINNET)
     const beraPoint = getAssetPoint(EndpointId.BERA_V2_MAINNET)
-    const cronosPoint = getAssetPoint(EndpointId.CRONOSEVM_V2_MAINNET)
+    const cronosevmPoint = getAssetPoint(EndpointId.CRONOSEVM_V2_MAINNET)
     const degenPoint = getAssetPoint(EndpointId.DEGEN_V2_MAINNET)
     const ethPoint = getAssetPoint(EndpointId.ETHEREUM_V2_MAINNET)
     const flarePoint = getAssetPoint(EndpointId.FLARE_V2_MAINNET)
@@ -57,7 +57,7 @@ export default async (): Promise<OmniGraphHardhat<AssetNodeConfig, AssetEdgeConf
     const arbContract = await getAssetNode(arbPoint)
     const baseContract = await getAssetNode(basePoint)
     const beraContract = await getAssetNode(beraPoint)
-    const cronosContract = await getAssetNode(cronosPoint)
+    const cronosevmContract = await getAssetNode(cronosevmPoint)
     const degenContract = await getAssetNode(degenPoint)
     const ethContract = await getAssetNode(ethPoint)
     const flareContract = await getAssetNode(flarePoint)
@@ -93,7 +93,7 @@ export default async (): Promise<OmniGraphHardhat<AssetNodeConfig, AssetEdgeConf
             ethContract,
             baseContract,
             beraContract,
-            cronosContract,
+            cronosevmContract,
             degenContract,
             flareContract,
             flowContract,
@@ -127,7 +127,7 @@ export default async (): Promise<OmniGraphHardhat<AssetNodeConfig, AssetEdgeConf
             ethPoint,
             basePoint,
             beraPoint,
-            cronosPoint,
+            cronosevmPoint,
             degenPoint,
             iotaPoint,
             islanderPoint,