+++ packages/stg-evm-v2/devtools/config/mainnet/01/feelib-v1.usdc.config.ts
@@ -15,7 +15,7 @@ import {
     onBsc,
     onCodex,
     onCoredao,
-    onCronos,
+    onCronosevm,
     onDegen,
     onEth,
     onFlare,
@@ -69,7 +69,7 @@ export default async (): Promise<OmniGraphHardhat<FeeLibV1NodeConfig, FeeLibV1Ed
     const bscFeeLibV1 = onBsc(contract)
     const codexFeeLibV1 = onCodex(contract)
     const coredaoFeeLibV1 = onCoredao(contract)
-    const cronosFeeLibV1 = onCronos(contract)
+    const cronosevmFeeLibV1 = onCronosevm(contract)
     const degenFeeLibV1 = onDegen(contract)
     const ethFeeLibV1 = onEth(contract)
     const flareFeeLibV1 = onFlare(contract)
@@ -145,7 +145,7 @@ export default async (): Promise<OmniGraphHardhat<FeeLibV1NodeConfig, FeeLibV1Ed
                 config: defaultNodeConfig,
             },
             {
-                contract: cronosFeeLibV1,
+                contract: cronosevmFeeLibV1,
                 config: defaultNodeConfig,
             },
             {