+++ packages/stg-evm-v2/devtools/config/mainnet/utils.ts
@@ -13,7 +13,7 @@ export const onBlast = withEid(EndpointId.BLAST_V2_MAINNET)
 export const onBsc = withEid(EndpointId.BSC_V2_MAINNET)
 export const onCodex = withEid(EndpointId.CODEX_V2_MAINNET)
 export const onCoredao = withEid(EndpointId.COREDAO_V2_MAINNET)
-export const onCronos = withEid(EndpointId.CRONOSEVM_V2_MAINNET)
+export const onCronosevm = withEid(EndpointId.CRONOSEVM_V2_MAINNET)
 export const onDegen = withEid(EndpointId.DEGEN_V2_MAINNET)
 export const onEth = withEid(EndpointId.ETHEREUM_V2_MAINNET)
 export const onEtherLink = withEid(EndpointId.ETHERLINK_V2_MAINNET)
@@ -75,7 +75,7 @@ export const chainFunctions = {
     'bsc-mainnet': onBsc,
     'codex-mainnet': onCodex,
     'coredao-mainnet': onCoredao,
-    'cronos-mainnet': onCronos,
+    'cronosevm-mainnet': onCronosevm,
     'degen-mainnet': onDegen,
     'ethereum-mainnet': onEth,
     'etherlink-mainnet': onEtherLink,
@@ -135,7 +135,7 @@ export const validCreditMessagingChains = new Set([
     'bsc-mainnet',
     'codex-mainnet',
     'coredao-mainnet',
-    'cronos-mainnet',
+    'cronosevm-mainnet',
     'degen-mainnet',
     'ethereum-mainnet',
     'flare-mainnet',
@@ -185,7 +185,7 @@ export const validTokenMessagingChains = new Set([
     'bsc-mainnet',
     'codex-mainnet',
     'coredao-mainnet',
-    'cronos-mainnet',
+    'cronosevm-mainnet',
     'degen-mainnet',
     'ethereum-mainnet',
     'flare-mainnet',